Moved auth files to its own model folder
